Committing full dev dataset evaluator
Committed an evaluator that allows the user to choose a desired amount of questions to evaluate from the full dev dataset. Most of the code is pulled from the evaluator provided to us in the notebook. As of right now the agent only uses the CoT algorithm when being evaluated.